reorganise selected file options
